fix(usage): SUM over a bigint column is NUMERIC, and sqlx will not read it as i64 — cast every such aggregate ::bigint at the SQL boundary
The Billing page's Usage section rendered "mismatched types; Rust type
`i64` (as SQL type `INT8`) is not compatible with SQL type `NUMERIC`"
verbatim. GET /api/usage/summary reads app_ai_calls through four
aggregate queries whose SUM(prompt_tokens|completion_tokens|cost_micros)
widen to NUMERIC in Postgres; the query parses and runs, then fails at
the first row decode. cargo check cannot see it (sqlx::query_as is
untyped), and an empty table does not dodge it either — the aggregate's
type is fixed at plan time, so the endpoint 500'd on a fresh box too.

Fixed at the boundary, not by swallowing (CLAUDE.md): the aggregate is
cast ::bigint so the column type matches the struct. Same class, same
fix, in every sibling that sums a bigint column:

- api/ai_calls.rs: spend_by_feature, spend_by_model, spend_by_day,
  wallet_spend_between (the Usage tab breakdown, the 14-day chart, and
  the last-month-to-date comparison)
- api/chat_usage.rs: the five token SUMs behind GET /api/chats/:id/usage.
  These were read with Row::get, which panics on a decode error rather
  than returning it, so the GROUP BY fix in e489673d had unmasked this
  one. estimated_cost_usd is float8 and needs no cast.
- applet_runner/limits.rs: spend_micros_last_day and
  spend_micros_for_run — the applet spend gate would have errored on its
  first check instead of enforcing anything.
- api/drive.rs: reconcile_usage's SUM(size_bytes). The two CASE sums are
  over int literals and were already bigint.

Left alone, checked: SUM(step_count) in day_summary.rs and
dayline/context.rs (integer column, sums to bigint), and the sites in
scheduler/applets.rs, lake.rs and lifeline.rs that already cast.

Regression test in ai_calls.rs runs all four readers on an empty scratch
database and again after a wallet row and a BYO row, asserting the token
totals, the wallet-only cost, and the byo_calls count.

Verified: cargo test -p virtues --lib (592 passed, 0 failed, 17 ignored),
and the cast queries run against the dev database return bigint with the
real totals. The running dev core on :8000 was not restarted (it does
not rebuild on edit), so the Billing page shows the old error until it is.

+++ virtues-core/src/api/chat_usage.rs
@@ -280,11 +280,15 @@ pub async fn get_chat_usage(pool: &PgPool, chat_id: String) -> Result<ChatUsageI
     let usage_row = sqlx::query(
         r#"
         SELECT
-            COALESCE(SUM(input_tokens), 0) as "input_tokens",
-            COALESCE(SUM(output_tokens), 0) as "output_tokens",
-            COALESCE(SUM(reasoning_tokens), 0) as "reasoning_tokens",
-            COALESCE(SUM(cache_read_tokens), 0) as "cache_read_tokens",
-            COALESCE(SUM(cache_write_tokens), 0) as "cache_write_tokens",
+            -- `::bigint` on each SUM: Postgres widens SUM(bigint) to NUMERIC,
+            -- which sqlx will not decode as i64 — and `get` below panics on
+            -- a decode error rather than returning it. `estimated_cost_usd`
+            -- is float8 and sums to float8, so it needs no cast.
+            COALESCE(SUM(input_tokens), 0)::bigint as "input_tokens",
+            COALESCE(SUM(output_tokens), 0)::bigint as "output_tokens",
+            COALESCE(SUM(reasoning_tokens), 0)::bigint as "reasoning_tokens",
+            COALESCE(SUM(cache_read_tokens), 0)::bigint as "cache_read_tokens",
+            COALESCE(SUM(cache_write_tokens), 0)::bigint as "cache_write_tokens",
             COALESCE(SUM(estimated_cost_usd), 0.0) as "total_cost"
         FROM app_chat_usage
         WHERE chat_id = $1
